Split project into Gradle modules: server and webclient
Server code has been moved to the "server" directory.
Each module has its own build file so that Angular tasks are grouped together.
