+++ server/src/main/kotlin/com/github/thibseisel/music/ErrorPayload.kt
