+++ server/src/main/kotlin/com/github/thibseisel/music/GeneralExceptionController.kt
