+++ server/src/main/kotlin/com/github/thibseisel/music/MainPageRedirect.kt
