+++ server/src/main/kotlin/com/github/thibseisel/music/SearchResult.kt
