+++ server/src/main/kotlin/com/github/thibseisel/music/spotify/SpotifyAlbum.kt
