+++ server/src/main/kotlin/com/github/thibseisel/music/spotify/SpotifyEntity.kt
