+++ server/src/main/kotlin/com/github/thibseisel/music/spotify/SpotifySearchResult.kt
