+++ server/src/main/resources/public/error/404.html
