+++ server/src/main/resources/public/error/500.html
