+++ server/src/main/resources/public/errors.css
