+++ settings.gradle.kts
@@ -1,1 +1,3 @@
 rootProject.name = "odeon-web"
+include(":webclient")
+include(":server")
